requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiohttp>=3.9.1
markdownify>=0.14.1